requirement.txt
asgiref==3.2.5                                                                                                          
bcrypt==3.1.7                                                                                                           
certifi==2019.11.28                                                                                                     
cffi==1.14.0                                                                                                            
configparser==5.0.0                                                                                                     
                                                                                                         
django-autoslug-iplweb==1.9.4.dev0                                                                                      
django-cities-light==3.5.0                                                                                              
django-xhtml2pdf==0.0.4                                                                                                 
html5lib==1.1                                                                                                                                                                                                                                                                                                                           
sweetify==2.0.0  
Pillow==7.1.1
progressbar2==3.50.1                                                                                                    
pycparser==2.20                                                                                                         
PyMySQL==0.9.3                                                                                                          
PyPDF2==1.26.0                                                                                                          
python-dateutil==2.8.1                                                                                                  
python-utils==2.4.0                                                                                                     
pytz==2019.3                                                                                                            
reportlab==3.5.42                                                                                                       
six==1.14.0                                                                                                             
sqlparse==0.3.1                                                                                                         
sweetify==2.0.0                                                                                                         
Unidecode==1.1.1                                                                                                        
webencodings==0.5.1                                                                                                     
xhtml2pdf==0.2.4
xlwt==1.3.0 